Enhance AppImage handling: update ToByteSource method to support async operations, refactor Runtime class to include Architecture, and improve AppImageFactory to manage application files more effectively.
