+++ src/DotnetPackaging.AppImage/Core/AppImageExtensions.cs
@@ -5,10 +5,11 @@ namespace DotnetPackaging.AppImage.Core;
 
 public static class AppImageExtensions
 {
-    public static Result<IByteSource> ToByteSource(this AppImageContainer appImageContainer)
+    public static Task<Result<IByteSource>> ToByteSource(this AppImageContainer appImageContainer)
     {
-        return SquashFS.Create(appImageContainer.Container).Map(sqfs =>
+        return SquashFS.Create(appImageContainer.Container).Map(async sqfs =>
         {
+            await sqfs.WriteTo($"/home/jmn/Escritorio/AppImageContents-{appImageContainer.Runtime.Architecture}.squashfs");
             var appImageBytes = appImageContainer.Runtime.Concat(sqfs);
             return ByteSource.FromByteChunks(appImageBytes);
         });
