+++ src/DotnetPackaging.AppImage/Core/Runtime.cs
@@ -2,9 +2,10 @@ using Zafiro.DivineBytes;
 
 namespace DotnetPackaging.AppImage.Core;
 
-public class Runtime(IByteSource source) : IRuntime
+public class Runtime(IByteSource source, Architecture architecture) : IRuntime
 {
     public IByteSource Source { get; } = source;
+    public Architecture Architecture { get; } = architecture;
 
     public IDisposable Subscribe(IObserver<byte[]> observer)
     {
